Splitting official tests of kernel and benchmarking
